requirements.txt
asgiref==3.5.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.13
django-allauth==0.51.0
django-debug-toolbar==3.4.0
django-etc==1.3.1
django-heroku==0.3.1
django-hitcount==1.3.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
gunicorn==20.1.0
idna==3.3
lxml==4.9.0
oauthlib==3.2.1
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
types-cryptography==3.3.23
typing_extensions==4.2.0
urllib3==1.26.12
whitenoise==6.1.0

httpx~=0.23.0
flake8==6.0.0